+++ tests/integ/modin/window/test_expanding.py
@@ -4,22 +4,115 @@
 
 import modin.pandas as pd
 import numpy as np
+import pandas as native_pd
 import pytest
 
+import snowflake.snowpark.modin.plugin  # noqa: F401
+from snowflake.snowpark.modin.plugin._internal.window_utils import (
+    IMPLEMENTED_EXPANDING_AGG_FUNCS,
+)
 from tests.integ.modin.sql_counter import sql_count_checker
+from tests.integ.modin.utils import eval_snowpark_pandas_result
+
+agg_func = pytest.mark.parametrize("agg_func", IMPLEMENTED_EXPANDING_AGG_FUNCS)
+min_periods = pytest.mark.parametrize("min_periods", [None, 0, 1, 2, 10])
+
+
+@agg_func
+@min_periods
+@sql_count_checker(query_count=1)
+def test_expanding_dataframe(agg_func, min_periods):
+    native_df = native_pd.DataFrame(
+        {"A": ["h", "e", "l", "l", "o"], "B": [0, -1, 2.5, np.nan, 4]}
+    )
+    snow_df = pd.DataFrame(native_df)
+    eval_snowpark_pandas_result(
+        snow_df,
+        native_df,
+        lambda df: getattr(
+            df.expanding(min_periods),
+            agg_func,
+        )(numeric_only=True),
+    )
+
+
+@agg_func
+@min_periods
+@sql_count_checker(query_count=1)
+def test_expanding_null_dataframe(agg_func, min_periods):
+    native_df = native_pd.DataFrame(
+        {
+            "A": ["h", np.nan, "l", "l", "o"],
+            "B": [np.nan, np.nan, np.nan, np.nan, np.nan],
+        }
+    )
+    snow_df = pd.DataFrame(native_df)
+    eval_snowpark_pandas_result(
+        snow_df,
+        native_df,
+        lambda df: getattr(
+            df.expanding(min_periods),
+            agg_func,
+        )(numeric_only=True),
+    )
+
+
+@agg_func
+@min_periods
+@sql_count_checker(query_count=1)
+def test_expanding_series(agg_func, min_periods):
+    native_series = native_pd.Series([0, -1, 2.5, np.nan, 4])
+    snow_series = pd.Series(native_series)
+    eval_snowpark_pandas_result(
+        snow_series,
+        native_series,
+        lambda df: getattr(
+            df.expanding(min_periods),
+            agg_func,
+        )(),
+    )
+
+
+@sql_count_checker(query_count=1)
+def test_expanding_min_periods_default():
+    native_df = native_pd.DataFrame(
+        {"A": ["h", "e", "l", "l", "o"], "B": [0, -1, 2.5, np.nan, 4]}
+    )
+    snow_df = pd.DataFrame(native_df)
+    eval_snowpark_pandas_result(
+        snow_df,
+        native_df,
+        lambda df: df.expanding().min(numeric_only=True),
+    )
+
+
+@sql_count_checker(query_count=0)
+def test_expanding_min_periods_negative():
+    native_df = native_pd.DataFrame({"B": [0, 1, 2, np.nan, 4]})
+    snow_df = pd.DataFrame(native_df)
+    eval_snowpark_pandas_result(
+        snow_df,
+        native_df,
+        lambda df: df.expanding("invalid_value").min(),
+        expect_exception=True,
+        expect_exception_type=ValueError,
+        expect_exception_match="min_periods must be an integer",
+    )
+    eval_snowpark_pandas_result(
+        snow_df,
+        native_df,
+        lambda df: df.expanding(-2).min(),
+        expect_exception=True,
+        expect_exception_type=ValueError,
+        expect_exception_match="min_periods must be >= 0",
+    )
 
 
 @pytest.mark.parametrize(
     "agg_func, agg_func_kwargs",
     [
         ("count", None),
-        ("sum", None),
-        ("mean", None),
         ("median", None),
-        ("var", 1),
-        ("std", 1),
-        ("min", None),
-        ("max", None),
         ("corr", None),
         ("cov", None),
         ("skew", None),
@@ -42,13 +135,7 @@ def test_expanding_aggregation_dataframe_unsupported(agg_func, agg_func_kwargs):
     "agg_func, agg_func_kwargs",
     [
         ("count", None),
-        ("sum", None),
-        ("mean", None),
         ("median", None),
-        ("var", 1),
-        ("std", 1),
-        ("min", None),
-        ("max", None),
         ("corr", None),
         ("cov", None),
         ("skew", None),